Add all options required to manage system users
For my use cases I need to be able to manage system accounts (specifically
accounts that will be created by future installs) which means setting custom
UIDs and home directories.

This change adds those and various other options which are useful when
attempting to work with users in non default configurations.